# openapi.yaml
openapi: 3.0.3
info:
  title: LIS Labo-Anapath API
  description: API REST de gestion de laboratoire d'anatomopathologie
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Développement local
  - url: https://api.labo-anapath.com/api/v1
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      tags: [Users]
      summary: Liste des utilisateurs
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - $ref: '#/components/parameters/search'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPageApiResponse'
    post:
      tags: [Users]
      summary: Créer un utilisateur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: Créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserApiResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/Conflict'
  /users/{id}:
    parameters:
      - $ref: '#/components/parameters/id'
    get:
      tags: [Users]
      summary: Détail d'un utilisateur
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserApiResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Users]
      summary: Modifier un utilisateur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserApiResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Users]
      summary: Supprimer un utilisateur
      responses:
        '200':
          $ref: '#/components/responses/EmptySuccess'
  /users/{id}/password:
    parameters:
      - $ref: '#/components/parameters/id'
    patch:
      tags: [Users]
      summary: Changer le mot de passe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordRequest'
      responses:
        '200':
          $ref: '#/components/responses/EmptySuccess'
        '400':
          $ref: '#/components/responses/ValidationError'

  # ── ROLES ───────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── REUSABLE PARAMETERS ──────────────────────────────────────────────────────
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    page:
      name: page
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
    size:
      name: size
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    search:
      name: search
      in: query
      schema:
        type: string
...
  responses:
    EmptySuccess:
      description: Opération réussie
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/EmptyApiResponse'
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorApiResponse'
...
    Conflict:
      description: Ressource déjà existante
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorApiResponse'
...
    ValidationError:
      description: Données de saisie invalides
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorApiResponse'

  # ── SCHEMAS ──────────────────────────────────────────────────────────────────
  schemas:
...
    PageMeta:
      type: object
      properties:
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 20
        totalElements:
          type: integer
          format: int64
          example: 100
        totalPages:
          type: integer
          example: 5
        last:
          type: boolean
    EmptyApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: OK
        data:
          nullable: true
        timestamp:
          type: string
          format: date-time
    ErrorApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Ressource introuvable
        data:
          nullable: true
        timestamp:
          type: string
          format: date-time
    ValidationErrorApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Erreurs de validation
        data:
          type: object
          additionalProperties:
            type: string
          example:
            firstName: Le prénom est obligatoire
            email: L'email doit être valide
        timestamp:
          type: string
          format: date-time
...
    UpdatePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        isActive:
          type: boolean
        branchId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
...
    UserRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        phone:
          type: string
        isActive:
          type: boolean
          default: true
    UserApiResponse:
      type: object
      properties:
        success: {type: boolean}
        message: {type: string}
        data:
          $ref: '#/components/schemas/UserResponse'
        timestamp: {type: string, format: date-time}
    UserPageApiResponse:
      type: object
      properties:
        success: {type: boolean}
        message: {type: string}
        data:
          allOf:
            - $ref: '#/components/schemas/PageMeta'
            - type: object
              properties:
                content:
                  type: array
                  items:
                    $ref: '#/components/schemas/UserResponse'
        timestamp: {type: string, format: date-time}

    # -- Role / Permission ----------------------------------------------------
...